[#67] refactor: index.ts 이름 변경
